Reorder providers according to pattern specificity
I saw a few occurrences where a Doubleclick request was being counted as
a Universal Analytics request, because the very simple UA pattern was
matching the Doubleclick URL.

To avoid this, I moved the providers with patterns using full domain
names to the top of the provider list, which means they'll be tested and
matched first.  The simpler, path-only patterns were moved to the
bottom.

Also updated the Doubleclick pattern to catch a few more cases.